Move magnetic moment helpers to PBC tools
