+++ gpu4pyscf/pbc/tools/__init__.py
@@ -13,3 +13,4 @@
 # limitations under the License.
 
 from .pbc import *
+from .magmom import *
